backend/core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["nnnLik <rslmakhmudov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "==5.0.7"
pydantic-settings = "==2.2.1"
psycopg2-binary = "==2.9.9"
django-cors-headers = "==4.3.1"
gunicorn = "==21.2.0"
pillow = "==10.3.0"
requests = "==2.31.0"
drf-yasg = "==1.21.7"
djangorestframework-simplejwt = "==5.3.1"
djangorestframework = "==3.15.2"
djoser = "==2.2.3"
django-storages = "==1.14.4"
boto3 = "==1.34.153"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
django-debug-toolbar = "^4.4.2"
ipython = "^8.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
